+++ app/languages/en_US.js
@@ -396,7 +396,7 @@ function en_USLang() {
     STR_APP_ANIMATIONS = "Enable app animations";
     STR_APP_ANIMATIONS_SUMMARY = "Controls side panel, scroll, notification and related animations";
     STR_UI_SETTINGS = "Interface customization's, color style, animations and related";
-    STR_GENERAL_CUSTOM = "Content customizations, sorting, auto refresh, timeouts and related";
+    STR_GENERAL_CUSTOM = "Content customization's, sorting, auto refresh, timeouts and related";
     STR_RUNNINGTIME = "App running for:";
     STR_410_ERROR = "Unable to get video link";
     STR_410_FEATURING = "Third party app are current without access for this featuring.";